# Ques_backend/docs/api/api_specification.yaml
openapi: 3.0.3
info:
  title: Ques Backend API
  description: |
    Complete API specification for the Ques application backend.
    Includes AI-powered user discovery, chat system, swipe mechanics, payments, and more.
  version: 1.0.0
  contact:
    name: Ques Backend Team
    email: backend@ques.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ques.app
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        bio:
          type: string
        phone_number:
          type: string
        wechat_id:
          type: string
        profile_picture_url:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - Users
      summary: Update user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: Profile updated
...
  /api/v1/users/search:
    get:
      tags:
        - Users
      summary: Search users
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: array
            items:
              type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results

  # Whisper endpoints
...
  /whispers:
    get:
      tags:
        - Whispers
      summary: Get received whispers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, declined, expired]
      responses:
        '200':
          description: Whispers retrieved
...
  /whispers/settings:
    get:
      tags:
        - Whispers
      summary: Get whisper settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved

    put:
      tags:
        - Whispers
      summary: Update whisper settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wechatId:
                  type: string
                customMessage:
                  type: string
                autoAccept:
                  type: boolean
      responses:
        '200':
          description: Settings updated

  # Payment endpoints
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Notifications retrieved

  # Contact endpoints
  /contacts:
    get:
      tags:
        - Contacts
      summary: Get contact list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Contacts retrieved

    post:
      tags:
        - Contacts
      summary: Add contact
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contactId:
                  type: string
                notes:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Contact added

  # User Settings endpoints
  /user-settings:
    get:
      tags:
        - User Settings
      summary: Get user settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved

    put:
      tags:
        - User Settings
      summary: Update user settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notifications:
                  type: object
                preferences:
                  type: object
      responses:
        '200':
          description: Settings updated

  # AI Services endpoints
...